pyproject.toml
[project]
name = "voicechat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "setuptools<81",
    "soundfile>=0.13.1",
    "SpeechRecognition>=3.10.4",
    "vosk>=0.3.45",
    "webrtcvad>=2.0.10",
]

[project.optional-dependencies]
wake = [
    "pvporcupine",
    "silero-vad",
]
google_stt = [
    "google-cloud-speech>=2.27.0",
]

[tool.uv.workspace]
members = [
    "alexa",
]
